Refactor: move exceptions to dedicated module and blake3 to hash.py
Create lib/cfv/exceptions.py module containing all exception classes
(FilenameError, CFVException, MissingDependencyError, CFVValueError,
CFVNameError, CFVSyntaxError, CFError). This avoids circular imports
when hash.py needs to raise MissingDependencyError.

Move blake3 hashing code (_blake3_module, _get_blake3, getfileblake3)
from common.py to hash.py where other hash functions reside.
